Dashboard/api/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
paho-mqtt>=2.1.0
mysql-connector-python>=9.0.0
gevent>=24.2.1
